// resources/ts/app/package/modules/spinner/types.ts
/*****************************************
 * Package Module Spinner
 *
 * Types
 *****************************************/

/*----------------------------------------*
 * Spinner Text
 *----------------------------------------*/

/**
 * Spinner Text
 */
export type SpinnerText =
    | "waiting"
    | "loading"
    | "processing"
    | "saving"
    | "deleting"
    | "fetching"
    | "searching"
    | "submitting";